Began fleshing out "playlist" feature. - The idea is that a user can add/remove/swap and save tracks in their playlists if they so choose to.
Poster is now separately defined as a custom HTML element.
- Posters have buttons to let users start playback immediately, add the
folder/video to the current playlist and an info button for showing
a brief description of the selection.

Included "Music" as a link and changed the nomenclature around
"Video" to "Media."

Completely separated the player from the controller.
- Is this a good idea? I don't know but it seems to be okay, for now.

Most of the playlist logic is completed
- Modified playTrack() to accept the id of the desired track to be played instead of having to do this difference math.
- Added icons to the playlist tracks.
- Added active classes to show which playlist element is selected.
- Had to rename some variables to be more concise as to which DOM element is being targeted.
** UI imcomplete **

UI work for playlist viewer.
- Video attaches and detaches when viewer is toggled.
- Corrected close button style and position.
- Media controls persist throughout DOM.

Updated readme.

The back button is working now
- A cache of visited pages is kept so you can navigate the directories easier.